auth_bot/requirements.txt
# Authorization Bot Requirements

# Telegram Bot Framework
python-telegram-bot>=20.0

# Database
pymongo>=4.0.0
dnspython>=2.0.0

# Environment Variables
python-dotenv>=0.19.0

# Cryptography
cryptography>=36.0.0

# JWT Token Handling
PyJWT>=2.6.0

# HTTP Requests
requests>=2.27.0

# Date/Time Handling
pytz>=2021.3

# Payment Processing
stripe>=4.0.0
paypalrestsdk>=1.13.1
razorpay>=1.3.0

# Logging
colorlog>=6.6.0

# Async HTTP Client (for analytics dashboard)
aiohttp>=3.8.0

# Async MongoDB Driver (for database handler)
motor>=3.0.0